Add TODO to move GameDescription to common package
This type is serialized between lobby nodes, and thus cannot be moved
until the next incompatible release.  However, the associated unit tests
were moved at this time to signal this intent.
